+++ dev-packages/node-integration-tests/package.json
@@ -25,14 +25,14 @@
   "dependencies": {
     "@anthropic-ai/sdk": "0.63.0",
     "@apollo/server": "^5.5.0",
-    "@aws-sdk/client-dynamodb": "3.1041.0",
-    "@aws-sdk/client-kinesis": "3.1041.0",
-    "@aws-sdk/client-lambda": "3.1041.0",
-    "@aws-sdk/client-s3": "3.1041.0",
-    "@aws-sdk/client-secrets-manager": "3.1041.0",
-    "@aws-sdk/client-sfn": "3.1041.0",
-    "@aws-sdk/client-sns": "3.1041.0",
-    "@aws-sdk/client-sqs": "3.1041.0",
+    "@aws-sdk/client-dynamodb": "^3.1046.0",
+    "@aws-sdk/client-kinesis": "^3.1046.0",
+    "@aws-sdk/client-lambda": "^3.1046.0",
+    "@aws-sdk/client-s3": "^3.1046.0",
+    "@aws-sdk/client-secrets-manager": "^3.1046.0",
+    "@aws-sdk/client-sfn": "^3.1046.0",
+    "@aws-sdk/client-sns": "^3.1046.0",
+    "@aws-sdk/client-sqs": "^3.1046.0",
     "@google/genai": "^1.20.0",
     "@growthbook/growthbook": "^1.6.1",
     "@hapi/hapi": "^21.3.10",
